sprint 1 done

+++ src/main/java/com/ust/pos/user/UserController.java
@@ -1,8 +1,11 @@
 package com.ust.pos.user;
 
 import com.ust.pos.dto.UserDto;
+import com.ust.pos.role.service.RoleService;
 import com.ust.pos.user.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -10,6 +13,8 @@ import org.springframework.web.bind.annotation.*;
 @Controller
 @RequestMapping("/user")
 public class UserController {
+    @Autowired
+    private RoleService roleService;
 
     @Autowired
     private UserService userService;
@@ -21,8 +26,9 @@ public class UserController {
     }
 
     @GetMapping("/get")
-    public String update(Model model, @RequestParam String username) {
+    public String update(Model model, @RequestParam String username, @ModelAttribute UserDto user) {
         UserDto response = userService.findByUserName(username);
+        model.addAttribute("roles", roleService.findAll());
         model.addAttribute("user", response);
         return "user/user";
     }
@@ -32,6 +38,7 @@ public class UserController {
         UserDto response = userService.update(userDto);
         if (!response.isSuccess()) {
             model.addAttribute("message", response.getMessage());
+            model.addAttribute("user", response);
             return "user/user";
         }
         return "redirect:/user/list";
@@ -39,7 +46,16 @@ public class UserController {
 
     @GetMapping("/delete")
     public String delete(Model model, @RequestParam String username) {
-        userService.delete(username);
-        return "user/user";
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(authentication != null){
+            String loggedInUser = authentication.getName();
+            userService.delete(username);
+            if (loggedInUser.equals(username)) {
+                SecurityContextHolder.clearContext();
+                return "redirect:/logout";
+            }
+            return "redirect:/user/list";
+        }
+        return null;
     }
 }
